requirements.txt
certifi>=2018.10.15
chardet>=3.0.4
python-dateutil>=2.8.0
Django==2.2
django-analytical>=2.5.0
django-debug-toolbar>=1.10.1
django-memcache-status>=1.3
idna>=2.7
pysolr>=3.8.1
python-memcached>=1.59
pytz>=2018.7
PyYAML>=3.13
requests>=2.21.0
simplejson>=3.16.0
six>=1.12
sqlparse>=0.2.4
urllib3>=1.24.1
